backend/requirement.txt
APScheduler==3.10.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.58.1
django-cors-headers==4.3.0
django-filter==23.3
django-mdeditor==0.1.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
feedparser==6.0.10
idna==3.4
Markdown==3.5
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
